requirements.txt
azure-functions==1.18.0
PyYAML==6.0.2
